Improved stereoscopic 3D settings

+++ Source/Core/VideoCommon/VideoConfig.h
@@ -310,12 +310,12 @@ struct VideoConfig final
   // Stereoscopy
   StereoMode stereo_mode{};
   bool stereo_per_eye_resolution_full = false;
-  int iStereoDepth = 0;
-  int iStereoConvergence = 0;
-  int iStereoConvergencePercentage = 0;
+  float iStereoDepth = 0;
+  float iStereoConvergence = 0;
+  float iStereoConvergencePercentage = 0;
   bool bStereoSwapEyes = false;
   bool bStereoEFBMonoDepth = false;
-  int iStereoDepthPercentage = 0;
+  float iStereoDepthPercentage = 0;
 
   // D3D only config, mostly to be merged into the above
   int iAdapter = 0;
